core/requirements.txt
sounddevice
scipy
numpy
pygame
google-generativeai
openai
gTTS
python-dotenv
urllib3<2
